refactor(storage): extract proximadb-storage-operations crate (#348)
Move src/storage/operations (storage maintenance: compaction [leveled/tiered]
plus scheduler, flush, operation locks, requantization, auto-scheduling) into a
dedicated workspace crate as a slice of the root-crate decomposition. The root
crate re-exports it as crate::storage::operations, so consumers are untouched.

Enabler: relocate the standalone corpus-version registry from the root crate
(src/catalog/corpus_version.rs, 667 LOC, zero outward crate deps) into the
proximadb-catalog crate, where corpus versioning architecturally belongs. The
root catalog module re-exports it (pub use proximadb_catalog::corpus_version),
so all 9 root consumers and the root FileSystemCorpusVersionStore are untouched.
This turns operations' compaction->version-bump call into a downward
storage->control dependency (validator-legal) instead of an upward edge to root.

Path rewrites in the moved code (no behavioral change):
- StorageEngineType: crate::storage::types shim -> proximadb_proto::proximadb_v1::StorageEngine
- CorpusVersionRegistry: crate::catalog -> proximadb_catalog (relocated module)
- internal absolute self-paths crate::storage::operations::X -> crate::X
- crate-level clippy allows (type_complexity/too_many_arguments/result_large_err)
  carried over from the root crate, under which this code was written

- root src/ shrinks ~6.1K LOC (operations ~5.4K + corpus_version 0.67K), two fewer modules
- proximadb-storage-operations: 58 tests pass; proximadb-catalog corpus_version: 25 pass
- clippy -D warnings clean (both crates, --lib gate); server/bench/migrate build
- layering green (storage tier; storage->control downward edge)

Ref ROOT_CRATE_DECOMPOSITION_PLAN_2026_06_21.
