+++ crates/control/proximadb-catalog/src/lib.rs
@@ -21,6 +21,10 @@ pub mod canonical_precision;
 // Collection-level DR / CRR engine contract (P1 of
 // COLLECTION_DR_CRR_ENGINE_CONTRACT.adoc).
 pub mod collection_dr_policy;
+// Global corpus-version registry + store trait (relocated from the root crate's
+// src/catalog so storage-side consumers like compaction can depend on it downward).
+pub mod corpus_version;
+pub use corpus_version::CorpusVersionRegistry;
 // Customer-facing DR policy mutation surface (S14 of the same contract).
 pub mod dr_policy_store;
 // DR reconciler decision logic (P3a of the same contract).
