+++ crates/storage/proximadb-storage-operations/Cargo.toml
@@ -1,0 +1,30 @@
+[package]
+name = "proximadb-storage-operations"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+authors.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "Storage maintenance operations for ProximaDB: compaction (leveled/tiered) + scheduler, flush, requantization, operation locks, and auto-scheduling"
+
+[lib]
+name = "proximadb_storage_operations"
+path = "src/lib.rs"
+
+[dependencies]
+anyhow.workspace = true
+async-trait.workspace = true
+chrono.workspace = true
+proximadb-catalog.workspace = true
+proximadb-proto.workspace = true
+proximadb-records.workspace = true
+proximadb-storage-common.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+tokio.workspace = true
+tracing.workspace = true
+
+[lints]
+workspace = true
